Add dimens file, splitted code into to files
