+++ app/src/main/java/com/example/dotaapp/UiElements.kt
@@ -12,8 +12,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.example.dotaapp.ui.theme.BlueCharcoal
-import com.example.dotaapp.ui.theme.PapayaWhip
 
 @Composable
 fun InstallButton(
@@ -35,6 +33,7 @@ fun InstallButton(
         )
     }
 }
+
 @Composable
 fun StarImage(
     isHalf: Boolean = false,
