+++ app/src/main/java/com/example/dotaapp/ui/theme/Theme.kt
@@ -1,7 +1,6 @@
 package com.example.dotaapp.ui.theme
 
 import android.app.Activity
-import android.view.Surface
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
