+++ app/src/main/java/com/example/dotaapp/ui/theme/Type.kt
@@ -2,7 +2,6 @@ package com.example.dotaapp.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
